requirements.txt
# Core Airflow installation
apache-airflow==2.6.2

# HTTP provider for Airflow (compatible version for Python 3.7)
apache-airflow-providers-http==4.4.1

# Requests library for making HTTP API calls
requests==2.28.1

# Pandas for data manipulation during the ETL process (compatible with Python 3.7)
pandas==1.3.5

# Environment variable management using python-dotenv
python-dotenv==0.21.1
